WS2-mvp/agent/requirements.txt
google-adk>=1.0.0
google-cloud-aiplatform>=1.60.0
google-genai>=1.0.0
uvicorn>=0.30.0
fastapi>=0.115.0
python-dotenv>=1.0.0
